+++ functional-tests/functional/dart/test/Nesting_test.dart
@@ -30,4 +30,15 @@ void main() {
 
     expect(result.first, 42);
   });
+  _testSuite.test("Call a method throwing an exception nested in a struct", () {
+    OuterStructInstantiationException exception = null;
+
+    try {
+      OuterStruct.doNothing();
+    } on OuterStructInstantiationException catch(e) {
+      exception = e;
+    }
+
+    expect(exception, isNotNull);
+  });
 }